Allow custom implementations for font data loading.
This adds a new public API, `GoogleFontsLoader`, and a corresponding new
instance field `fontLoader` in the global `GoogleFonts.config` class.

The new API allows users to hook the loading of fonts into the Flutter
engine, which is useful in scenarios where the actual font data is
needed for other purposes. The newly added default font loader simply
uses Flutter's `FontLoader` as before.

Fixes #102.